integrations/terra/requirements-full.txt
# GenomeVedic Python Library - Full Installation

# Core dependencies
-r requirements.txt

# Google Cloud Storage integration
google-cloud-storage>=2.10.0

# BAM file processing
pysam>=0.21.0

# Optional: Enhanced features
numpy>=1.20.0
pandas>=1.3.0
